[IPC] Separate IPC subcommands
